Issue #SBCOSS-734 fix: Refactor AssessmentItemActor to use AssessmentItemUtils
