+++ assessment-api/assessment-actors/src/main/scala/org/sunbird/actors/AssessmentItemActor.scala
@@ -14,6 +14,7 @@ import org.sunbird.utils.RequestUtil
 import org.sunbird.managers.AssessmentManager
 import org.sunbird.validators.AssessmentItemValidator
 import org.sunbird.utils.JavaJsonUtils
+import org.sunbird.utils.AssessmentItemUtils
 
 import java.util
 import javax.inject.Inject
@@ -34,46 +35,14 @@ class AssessmentItemActor @Inject()(implicit oec: OntologyEngineContext) extends
 
   def create(request: Request): Future[Response] = {
     val requestData = request.getRequest
-    val skipValidation = requestData.getOrDefault("skipValidation", false.asInstanceOf[AnyRef]).asInstanceOf[Boolean]
-    
-    val metadata = if (requestData.containsKey("metadata")) {
-      requestData.get("metadata").asInstanceOf[util.Map[String, AnyRef]]
-    } else {
-      throw new ClientException("ERR_ASSESSMENT_ITEM_CREATE", "Assessment Item metadata is missing")
-    }
-    
-    if (!metadata.containsKey("objectType")) {
-      metadata.put("objectType", "AssessmentItem")
-    }
-    
-    if (!metadata.containsKey("mimeType")) {
-      metadata.put("mimeType", "application/vnd.sunbird.assessmentitem")
-    }
-    
-    if (!metadata.containsKey("framework") || StringUtils.isBlank(metadata.get("framework").asInstanceOf[String])) {
-      metadata.put("framework", getDefaultFramework())
-    }
-
-    if (!metadata.containsKey("version")) {
-      metadata.put("version", java.lang.Integer.valueOf(1))
-    }
-    if (metadata.containsKey("level")) {
-      metadata.remove("level")
-    }
-
-    replaceMediaItemsWithVariants(metadata)
-    request.getRequest.remove("metadata")
-    request.getRequest.putAll(metadata)
-    
-    if (!request.getRequest.containsKey("objectType")) {
-      request.getRequest.put("objectType", "AssessmentItem")
-    }
-    
-    if (!skipValidation) {
-      AssessmentItemValidator.validateAssessmentItemRequest(request.getRequest, "ASSESSMENT_ITEM_CREATE")
-    }
+    val skipValidation = AssessmentItemUtils.getSkipValidation(requestData)
+    val metadata = AssessmentItemUtils.extractMetadata(requestData)
+    AssessmentItemUtils.populateDefaults(metadata)
+    AssessmentItemUtils.replaceMediaItemsWithVariants(metadata)
+    AssessmentItemUtils.flattenMetadataToRequest(request, metadata)
+    if (!skipValidation) AssessmentItemValidator.validateAssessmentItemRequest(requestData, "ASSESSMENT_ITEM_CREATE")
     DataNode.create(request).map { node =>
-      ResponseHandler.OK.put("identifier", node.getIdentifier.replace(".img", ""))
+      ResponseHandler.OK.put("identifier", node.getIdentifier.replace(".img", ""), "versionKey" -> node.getMetadata.get("versionKey"))
     }
   }
 
@@ -98,196 +67,58 @@ class AssessmentItemActor @Inject()(implicit oec: OntologyEngineContext) extends
   def update(request: Request): Future[Response] = {
     val requestData = request.getRequest
     request.getRequest.put("identifier", request.getContext.get("identifier"))
-
-    val skipValidation = requestData.getOrDefault("skipValidation", false.asInstanceOf[AnyRef]).asInstanceOf[Boolean]
-    
+    val skipValidation = AssessmentItemUtils.getSkipValidation(requestData)
     DataNode.read(request).flatMap(existingNode => {
       if (NodeUtil.isRetired(existingNode)) {
         throw new ClientException("ERR_ASSESSMENT_ITEM_UPDATE", "Cannot update retired assessment item: " + existingNode.getIdentifier)
       }
-      
       val metadata = if (requestData.containsKey("metadata")) {
         requestData.get("metadata").asInstanceOf[util.Map[String, AnyRef]]
       } else {
         throw new ClientException("ERR_ASSESSMENT_ITEM_UPDATE", "Assessment Item metadata is missing")
       }
-      
       if (!metadata.containsKey("framework")) {
         val existingFramework = existingNode.getMetadata.get("framework")
         if (existingFramework != null) {
           metadata.put("framework", existingFramework)
         }
       }
-
       if (metadata.containsKey("level")) {
         metadata.remove("level")
       }
-      
       if (!skipValidation) {
         AssessmentItemValidator.validateAssessmentItemRequest(requestData, "ASSESSMENT_ITEM_UPDATE")
       }
-      replaceMediaItemsWithVariants(metadata)
+      AssessmentItemUtils.replaceMediaItemsWithVariants(metadata)
       DataNode.update(request).map { node =>
-        ResponseHandler.OK.put("identifier", node.getIdentifier.replace(".img", ""))
+        ResponseHandler.OK.put("identifier", node.getIdentifier.replace(".img", ""), "versionKey" -> node.getMetadata.get("versionKey"))
       }
     })
   }
 
   def retire(request: Request): Future[Response] = {
     request.getRequest.put("identifier", request.getContext.get("identifier"))
-    
     DataNode.read(request).flatMap(node => {
       if (NodeUtil.isRetired(node)) {
         throw new ClientException("ERR_ASSESSMENT_ITEM_RETIRE", "Assessment Item is already retired: " + node.getIdentifier)
       }
-      validateRetirePermissions(request, node)
-      validateAssessmentItemUsage(node)
-      
+      AssessmentItemUtils.validateRetirePermissions(request, node)
+      AssessmentItemUtils.validateAssessmentItemUsage(node)
+      val identifier = request.get("identifier").asInstanceOf[String]
       val updateRequest = new Request(request)
-      updateRequest.put("status", "Retired")
-      updateRequest.put("lastStatusChangedOn", System.currentTimeMillis().toString)
-      updateRequest.put("lastUpdatedOn", System.currentTimeMillis().toString)
-      
-      DataNode.update(updateRequest).map(updatedNode => {
-        ResponseHandler.OK.put("identifier", updatedNode.getIdentifier.replace(".img", ""))
+      val identifiers = java.util.Arrays.asList(identifier, identifier + ".img")
+      updateRequest.put("identifiers", identifiers)
+      val date = Platform.getString("date.format", java.time.format.DateTimeFormatter.ISO_INSTANT.format(java.time.Instant.now()))
+      val updateMetadata: util.Map[String, AnyRef] = Map(
+        "prevStatus" -> node.getMetadata.get("status"),
+        "status" -> "Retired",
+        "lastStatusChangedOn" -> date,
+        "lastUpdatedOn" -> date
+      ).asJava
+      updateRequest.put("metadata", updateMetadata)
+      DataNode.bulkUpdate(updateRequest).map(_ => {
+        ResponseHandler.OK.put("identifier", node.getIdentifier.replace(".img", ""), "versionKey" -> node.getMetadata.get("versionKey"))
       })
     })
-  }
-
-  private def validateRetirePermissions(request: Request, node: Node): Unit = {
-    val currentStatus = node.getMetadata.getOrDefault("status", "Draft").asInstanceOf[String]
-    if (StringUtils.equalsIgnoreCase(currentStatus, "Processing")) {
-      throw new ClientException("ERR_ASSESSMENT_ITEM_RETIRE", "Cannot retire assessment item in processing state")
-    }
-  }
-
-  private def validateAssessmentItemUsage(node: Node): Unit = {
-    val inRelations = node.getInRelations
-    if (inRelations != null && !inRelations.isEmpty) {
-      val activeRelations = inRelations.asScala.filter(rel => 
-        rel.getStartNodeMetadata != null && 
-        !StringUtils.equalsIgnoreCase(rel.getStartNodeMetadata.getOrDefault("status", "").asInstanceOf[String], "Retired")
-      )
-      if (activeRelations.nonEmpty) {
-        TelemetryManager.warn("Assessment item has active relations but proceeding with retirement: " + node.getIdentifier)
-      }
-    }
-  }
-
-  
-  private def replaceMediaItemsWithVariants(assessmentItem: util.Map[String, AnyRef]): Unit = {
-    
-    val media = assessmentItem.get("media")
-    
-    if (media != null && StringUtils.isNotBlank(media.toString)) {
-      val mediaList = if (media.isInstanceOf[String]) {
-        JavaJsonUtils.deserialize[java.util.List[java.util.Map[String, Object]]](media.toString)
-      } else if (media.isInstanceOf[util.List[_]]) {
-        media.asInstanceOf[java.util.List[java.util.Map[String, Object]]]
-      } else {
-        null
-      }
-      
-      if (mediaList != null && !mediaList.isEmpty) {
-        var replaced = false
-        val resolution = Platform.getString("assessment.media.resolution", "low")
-        
-        val processedMediaList = mediaList.asScala.map { mediaItem =>
-          processMediaItem(mediaItem, resolution) match {
-            case Some(updatedItem) =>
-              replaced = true
-              updatedItem
-            case None => mediaItem
-          }
-        }.asJava
-        
-        if (replaced) {
-          val updatedMedia = JavaJsonUtils.serialize(processedMediaList)
-          assessmentItem.put("media", updatedMedia)
-        }
-      }
-    }
-  }
-  
-  private def processExternalMediaInContent(content: String): String = {
-    var processedContent = content
-    
-    val assetPattern = """asset_id['":\s]*([^'",\s}]+)""".r
-    assetPattern.findAllMatchIn(content).foreach { matchResult =>
-      val assetId = matchResult.group(1)
-      if (StringUtils.isNotBlank(assetId)) {
-        try {
-          val assetRequest = new Request()
-          assetRequest.getContext.put("identifier", assetId)
-          assetRequest.setOperation("getDataNode")
-
-          DataNode.read(assetRequest).map { assetNode =>
-            if (assetNode != null) {
-              val variantsJSON = assetNode.getMetadata.get("variants")
-              if (variantsJSON != null && StringUtils.isNotBlank(variantsJSON.toString)) {
-                val variants = JavaJsonUtils.deserialize[java.util.Map[String, String]](variantsJSON.toString)
-                if (variants != null && !variants.isEmpty) {
-                  val resolution = Platform.getString("assessment.media.resolution", "low")
-                  val variantURL = variants.get(resolution)
-                  if (StringUtils.isNotEmpty(variantURL)) {
-                    processedContent = processedContent.replace(assetId, variantURL)
-                  }
-                }
-              }
-            }
-          }
-        } catch {
-          case e: Exception =>
-            TelemetryManager.warn(s"Failed to process external media for asset: $assetId. Error: ${e.getMessage}")
-        }
-      }
-    }
-    
-    processedContent
-  }
-
-  private def processMediaItem(mediaItem: java.util.Map[String, Object], resolution: String): Option[java.util.Map[String, Object]] = {
-    var assetId = mediaItem.get("asset_id")
-    if (assetId == null) {
-      assetId = mediaItem.get("assetId")
-    }
-    
-    if (assetId != null && StringUtils.isNotBlank(assetId.toString)) {
-      val assetRequest = new Request()
-      assetRequest.getContext.put("identifier", assetId.toString)
-      assetRequest.setOperation("getDataNode")
-
-      DataNode.read(assetRequest).map { assetNode =>
-        if (assetNode != null) {
-        val variantsJSON = assetNode.getMetadata.get("variants")
-
-        if (variantsJSON != null && StringUtils.isNotBlank(variantsJSON.toString)) {
-          val variants = JavaJsonUtils.deserialize[java.util.Map[String, String]](variantsJSON.toString)
-
-          if (variants != null && !variants.isEmpty) {
-            val variantURL = variants.get(resolution)
-            if (StringUtils.isNotEmpty(variantURL)) {
-              val updatedMediaItem = new java.util.HashMap[String, Object](mediaItem)
-              updatedMediaItem.put("src", variantURL)
-
-              if (variants.containsKey("high")) {
-                updatedMediaItem.put("highResolutionSrc", variants.get("high"))
-              }
-              if (variants.containsKey("medium")) {
-                updatedMediaItem.put("mediumResolutionSrc", variants.get("medium"))
-              }
-
-              return Some(updatedMediaItem)
-            }
-          }
-        }
-      }
-      }
-    }
-    None
-  }
-
-  private def getDefaultFramework(): String = {
-    Platform.getString("assessment.default.framework", "NCF")
   }
 }
